Fix snake movement, eye positions, logic, incorrect variable usages, signness
Remove poor solution to network thread concurrency, to be resolved next

+++ src/main/java/net/gegy1000/slyther/client/ClientNetworkManager.java
@@ -80,7 +80,7 @@ public class ClientNetworkManager extends WebSocketClient {
             bytesPerSecond += buffer.length();
             client.lastPacketTime = client.currentPacketTime;
             client.currentPacketTime = System.currentTimeMillis();
-            short serverTimeDelta = buffer.readShort();
+            int serverTimeDelta = buffer.readShort();
             byte messageId = (byte) buffer.readByte();
             long timeDelta = client.currentPacketTime - client.lastPacketTime;
             if (client.lastPacketTime == 0) {
